+++ jgclark.PeriodicReviews/src/index.js
@@ -22,6 +22,8 @@ export {
   quarterlyReviewQuestions,
   yearlyReviewQuestions,
   onReviewWindowAction,
+  closePeriodicReviewWindow,
+  closePeriodicReviewWindowCommand,
 } from './periodReviews'
 
 export {